Extract email inline code handling to a concern
